Add files via upload
Data-Science-Numpy
NumPy stands for numeric python which is a python package for the computation and processing of the multidimensional and single dimensional array elements.It is an extension module of Python which is mostly written in C. It provides various functions which are capable of performing the numeric computations with a high speed.

WHY WE USE NUMPY ?

Numpy provides effienct storage. It provides better ways of handling data for processing NumPy arrays are faster and more compact than Python lists. It is easy to learn An array consumes less memory and is convenient to use. NumPy uses much less memory to store data and it provides a mechanism of specifying the data types. This allows the code to be optimized even further. It makes many mathematical operations used widely in scientific computing fast and easy to use, such as:

Vector-Vector multiplication Matrix-Matrix and Matrix-Vector multiplication Element-wise operations on vectors and matrices (i.e., adding, subtracting, multiplying, and dividing by a number ) Element-wise or array-wise comparisons Applying functions element-wise to a vector/matrix ( like pow, log, and exp) A whole lot of Linear Algebra operations can be found in NumPy.linalg Reduction, statistics, and much more